Disable e1000 driver in ARK
We tried to disable this driver in RHEL8, but the dev experience team
pushed back because virtbox only supports this driver.  We shouldn't
allow that to go on forever though, as the driver isn't being maintained
upstream anymore.  So lets turn it off in ARK and at least have the
conversation again in RHEL 9

Signed-off-by: Neil Horman <nhorman@redhat.com>