[redhat] New configs in drivers/dma-buf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DMABUF_MOVE_NOTIFY:

 Don''t pin buffers if the dynamic DMA-buf interface is available on both the
 exporter as well as the importer. This fixes a security problem where
 userspace is able to pin unrestricted amounts of memory through DMA-buf.
 But marked experimental because we don''t jet have a consistent execution
 context and memory management between drivers.

 Symbol: DMABUF_MOVE_NOTIFY [=n]
 Type  : bool
 Defined at drivers/dma-buf/Kconfig:42
   Prompt: Move notify between drivers (EXPERIMENTAL)
   Location:
     -> Device Drivers
       -> DMABUF options

Enabled at the request of Prarit Bhargava to avoid the security issue.

---

Cc: Prarit Bhargava <prarit@redhat.com>