Add overrides to filter-modules.sh
This adds an overrides list to filter-modules.sh and a section to
process this list.  It is to be used when a specific module which would
be filtered to kernel-modules needs to remain in kernel-core for depmod
reasons.  A current example of this is cec.ko. The drm_kms_helper.ko
module needs symbols from it if we enable CONFIG_DRM_DP_CEC.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>