New configs in net/bluetooth
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BT_MSFTEXT:

 This options enables support for the Microsoft defined HCI
 vendor extensions.

 Symbol: BT_MSFTEXT [=n]
 Type  : bool
 Defined at net/bluetooth/Kconfig:96
   Prompt: Enable Microsoft extensions
   Depends on: NET [=y] && BT [=m]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])

Cc: Gopal Tiwari <gtiwari@redhat.com>