redhat/Makefile.common: Fix MARKER
'make setup-source' is failing with, for example,

fatal: Not a valid object name master
BUILDID is "hello". Update '/home/prarit/git-kernel/kernel-ark/localversion' to change.
Gathering new log entries since 2eb3eaedb74aa011774bb95fd6516d6d7858d0f8
fatal: ambiguous argument 'itd3590ebf6f91.1-1-..': unknown revision or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'

As of 5.7.0 new tags were added to the tree of the form

	kernel-5.8.0-0.rc7.20200730gitd3590ebf6f91.1

After applying a patch the value of "git describe" is
kernel-5.8.0-0.rc7.20200730gitd3590ebf6f91.1-1-gd76dfb51aa86.  This string
is stored in TAG which is used by MARKER. The value of MARKER is set
using a cut command on 'g' which results in MARKER being set to
"it69119673bd50.1" instead of "d76dfb51aa86"

The problem with continuing to use cut and 'g' is that the tag string
may change again.  It is safer to use awk and '-g'.

Use awk instead of cut to evaluate MARKER.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>