[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_STATIC_CALL_SELFTEST:

 Boot time self-test of the call patching code.

 Symbol: STATIC_CALL_SELFTEST [=n]
 Type  : bool
 Defined at arch/Kconfig:109
   Prompt: Static call selftest
   Depends on: HAVE_STATIC_CALL [=y]
   Location:
     -> General architecture-dependent options

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>