redhat: handle certificate files conditionally as done for src.rpm
When we create the src.rpm, we don't need to copy all certificate files,
only a set of them are needed depending on the RELEASED_KERNEL status.
This is necessary because when you do a make dist-git, the script
rh-dist-git.sh will create a diff to compare what was copied against what
is on dist-git clone, and this report will be bogus if we copy uneeded
files (it'll show in the diff that some files are not in the dist-git
copy, since the src.rpm created doesn't contain all certificate files,
because it's conditionally added based on released_kernel value).

Also while at it, delete old certificate files (secureboot.cer,
securebootca.cer) which are not used anymore since last grub related
security update.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>