all: s390x: Increase CONFIG_PCI_NR_FUNCTIONS to 512 (#1888735)
CONFIG_PCI_NR_FUNCTIONS limits the number of PCI functions the kernel
can handle on s390. In some situations (e.g. with the ROCE Express
adapter in z15) the number of available functions can easily exceed the
the current maximum. Thus increase the limit to 512, which is also the
default upstream.

Signed-off-by: Dan Horák <dan@danny.cz>
CC: Philipp Rudo <prudo@redhat.com>
CC: Steve Best <sbest@redhat.com>