[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MOUSE_SERIAL:

 Say Y here if you have a serial (RS-232, COM port) mouse connected
 to your system. This includes Sun, MouseSystems, Microsoft,
 Logitech and all other compatible serial mice.

 If unsure, say N.

 To compile this driver as a module, choose M here: the
 module will be called sermouse.

 Symbol: MOUSE_SERIAL [=n]
 Type  : tristate
 Defined at drivers/input/mouse/Kconfig:203
   Prompt: Serial mouse
   Depends on: !UML && INPUT [=y] && INPUT_MOUSE [=y]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Mice (INPUT_MOUSE [=y])
 Selects: SERIO [=y]

---

Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>