Keep CONFIG_KASAN_HW_TAGS off for aarch64 debug configs
Some time ago this config option came in, and the scripts missed it. I
added the file to pending turning it off, but that doesn't get it
reviewed.  This MR would move the config from pending to common, but it
does keep CONFIG_KASAN_HW_TAGS off.  Let me know if you would rather
turn it on.

Cc: Prarit Bhargava <prarit@redhat.com>

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>