Fix ppc64le cross build packaging
CKI has been failing to cross build a ppc64le rpm for a while now. The
reason why is the ppc64le-gnu-strip command could not recognize the
scripts/dtc/libfdt/fdt_strerror.o file because it was an x86_64 file
type.  That is a result of the cross env.

What should have happened is, none of the *.o files in the script
directory should be there and thus the strip command would have nothing
to strip.

Having x86_64 files lying around to be packaged in kernel-devel is
another problem.  But for this build problem, just remove all the *.o
files in the script/ area.

The spec file tries to do this, but fails to remove *.o's at another
directory layer (libfdt).  Let's just replace the 'rm' commands with a
'find and rm' command instead.

Tested in a CKI container.  After this change, the rpm builds.

Signed-off-by: Don Zickus <dzickus@redhat.com>