Embed crypto algos, modes and templates needed in the FIPS mode
Currently a number of FIPS-allowed algorithms are built as modules or are
not enabled in Fedora and ARK. This can result in a panic while booting
in the FIPS mode. Fix this by embedding the FIPS-allowed algorithms, modes
and templates into a kernel, the same way as CTC, CBC and other algorithms
already do.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1947240
Signed-off-by: Vladis Dronov <vdronov@redhat.com>