redhat/configs: enable CONFIG_BLK_DEV_ZONED
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1638087

This option has been enabled in Fedora already.

Also it is mentioned in this BZ that 'our current plan is to enable this
in RHEL 9 first', so enable this option in redhat/configs/common so that
people can use ZNS via userspace or play it via nvme-cli.

Signed-off-by: Ming Lei <ming.lei@redhat.com>