[redhat] New configs in drivers/infiniband
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_INFINIBAND_IRDMA:

 This is an Intel(R) Ethernet Protocol Driver for RDMA driver
 that support E810 (iWARP/RoCE) and X722 (iWARP) network devices.

 Symbol: INFINIBAND_IRDMA [=n]
 Type  : tristate
 Defined at drivers/infiniband/hw/irdma/Kconfig:2
   Prompt: Intel(R) Ethernet Protocol Driver for RDMA
   Depends on: INFINIBAND [=m] && (INFINIBAND_USER_ACCESS [=m] || !INFINIBAND_USER_ACCESS [=m]) && INET [=y] && (IPV6 [=y] || !IPV6 [=y]) && PCI [=y] && ICE [=m] && I40E [=m]
   Location:
     -> Device Drivers
       -> InfiniBand support (INFINIBAND [=m])
 Selects: GENERIC_ALLOCATOR [=y] && CONFIG_AUXILIARY_BUS

---

Cc: Kamal Heib <kheib@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 CONFIG_INFINIBAND_IRDMA enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>