redhat/configs: Fix CONFIG_VIRTIO_IOMMU to 'y' on aarch64
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1972795
Upstream Status: RHEL only
Tested: on ARM with qemu (Jean-Philippe Brucker's qemu branch as the
qemu integration is not yet upstream, with both protected virtio-blk-pci
and a virtio-net-pci devices)

Commit fe6d9534dbe4 ("redhat/configs: Set CONFIG_VIRTIO_IOMMU on aarch64")
enabled the compilation of the virtio-iommu driver on ARM but set the
config to 'm'. However if there are devices behind the vIOMMU that are
needed during the boot such as a virtio-blk-pci device, the guest fails
to boot as the driver is not loaded. Let's fix this bootstrap issue
by compiling the driver in-kernel as it is done for other iommu drivers.

Signed-off-by: Eric Auger <eric.auger@redhat.com>