[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PERF_EVENTS_AMD_UNCORE [=y]
 Type  : tristate
 Defined at arch/x86/events/Kconfig:37
   Prompt: AMD Uncore performance events
   Depends on: PERF_EVENTS [=y] && CPU_SUP_AMD [=y]
   Location:
     -> Processor type and features
       -> Performance monitoring

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>