[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_RSMU_I2C [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:2202
   Prompt: Renesas Synchronization Management Unit with I2C
   Depends on: HAS_IOMEM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m]

---

 Symbol: MFD_RSMU_SPI [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:2216
   Prompt: Renesas Synchronization Management Unit with SPI
   Depends on: HAS_IOMEM [=y] && SPI [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
 Selects: MFD_CORE [=m] && REGMAP_SPI [=m]

---

Cc: Torez Smith <torez@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>