[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMMAND_LINE_SIZE [=4096]
 Type  : integer
 Range : [896 1048576]
 Defined at arch/s390/Kconfig:441
   Prompt: Maximum size of kernel command line
   Location:
     -> Processor type and features

---

 Symbol: S390_KPROBES_SANITY_TEST [=n]
 Type  : tristate
 Defined at arch/s390/Kconfig:963
   Prompt: Enable s390 specific kprobes tests
   Depends on: KPROBES [=y] && KUNIT [=m]
   Location:
     -> Selftests

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>