Fix up PREEMPT configs
While preempt_dynamic was greatly improved at the beginning of the merge
window, the config to make things work was messy and the new BEHAVIOUR
configs were confusing to some. With commit
a8b76910e465d718effce0cad306a21fa4f3526b the BEHAVIOUR config options
are gone and we now default to whichever CONFIG_PREEMPT is selected.
This MR moves us to the new config options, but should leave the kernel
behaving as before with PREEMPT_VOLUNTARY as the default with
PREEMPT_DYNAMIC enabled for all arches but s390 which does not enable
DYNAMIC and builds with PREEMPT_NONE.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>