redhat: add zstream switch for zstream release numbering
Bugzilla: INTERNAL
Upstream Status: RHEL only

When we fork from CentOS into RHEL stable releases, we need to switch the
release number to use zstream numbering, to not conflict with future
centos release numbers. We use the ZSTREAM=yes value in that case, keep
it set to no by default until we need it.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>