[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HID_PXRC [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:912
   Prompt: PhoenixRC HID Flight Controller
   Depends on: INPUT [=y] && HID [=y]
   Location:
     Main menu
       -> Device Drivers
         -> HID support
           -> HID bus support (HID [=y])
             -> Special HID drivers

---

 Symbol: HID_TOPRE [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:1099
   Prompt: Topre REALFORCE keyboards
   Depends on: INPUT [=y] && HID [=y]
   Location:
     Main menu
       -> Device Drivers
         -> HID support
           -> HID bus support (HID [=y])
             -> Special HID drivers

---

 Symbol: HID_VRC2 [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:457
   Prompt: VRC-2 Car Controller
   Depends on: INPUT [=y] && HID [=y]
   Location:
     Main menu
       -> Device Drivers
         -> HID support
           -> HID bus support (HID [=y])
             -> Special HID drivers

---

Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>