[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: QCOM_RAMP_CTRL [=n]
 Type  : tristate
 Defined at drivers/soc/qcom/Kconfig:117
   Prompt: Qualcomm Ramp Controller driver
   Depends on: ARCH_QCOM [=y] || COMPILE_TEST [=n]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Qualcomm SoC drivers
           -> Qualcomm Ramp Controller driver (QCOM_RAMP_CTRL [=n])

---

 Symbol: WPCM450_SOC [=n]
 Type  : tristate
 Defined at drivers/soc/nuvoton/Kconfig:2
   Prompt: Nuvoton WPCM450 SoC driver
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Nuvoton WPCM450 SoC driver (WPCM450_SOC [=n])
 Selects: SOC_BUS [=n]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>