xsk: Support UMEM chunk_size > PAGE_SIZE
Add basic support for UMEMs with chunk_size > PAGE_SIZE. For now,
restrict this to UMEMs backed by HugeTLB pages. This provides a way for
drivers to support AF_XDP with jumbo frames larger than PAGE_SIZE.
Initially, only XDP_COPY mode is usuable pending future driver work.

For simplicity, check for HugeTLB pages at UMEM registration. This
implies that hugepages are required for XDP_COPY mode despite the lack
of a DMA mapping. This restriction could be lifted in the future.

Even in HugeTLB mode, page accounting is done using order-0 (4 KiB)
pages. This minimizes the size of this change and reduces the risk of
impacting driver code.